ajuste na entidade conta, retorno do saldo para double

+++ bytebank/src/main/java/br/com/unpbankdigital/domain/entity/Conta.java
@@ -18,7 +18,7 @@ public class Conta {
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Long id;
     private Integer numero;
-    private Float saldo;
+    private Double saldo;
     @ManyToOne
     @JoinColumn(name = "titular_id")
     private Cliente titular;
